Final Project re-work (temp)
I reworked the program to include more user functionality, but there's a couple things I'm kinda caught up on. Besides not knowing how to access the json file you've provided me, Patrick, I can't figure out how to display the proper artist and track name just above the visualizer after selecting the artist and track. The specific issue is in my "itunesResults(json)" function (index.html file) where I take the data from the iTunes API and try to place the correct data into my div according to which track was selected. The selTrack variable I created doesn't seem to be showing any of the values in the trackList and I can't wrap my mind around why that is as it works fine for the artistList in my "artistSelect()" function.