Assignment 12 (Sophia's Project Edit)
I decided to edit Sophia's project. I couldn't get the code to work properly after submitting a key into the text box so I decided to instead give the frontend some fancy CSS. I also edited the script sources in the "frontend.html" file to no longer point to Sophia's file directory so it would be accessible to anyone with the project zip and I made a minor edit to the last read-out at the bottom of the frontend file to read "In the key of B..." so it wouldn't repeat the key of C again.